requirements.txt
torch
torchvision
timm
pillow
matplotlib
scikit-learn
gradio
grad-cam
ipykernel
